tests/requirements.txt
attrs==20.2.0
importlib-metadata==1.7.0
iniconfig==1.0.1
more-itertools==8.5.0
packaging==20.4
Pillow==7.2.0
pkg-resources==0.0.0
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==6.0.2
six==1.15.0
toml==0.10.1
zipp==3.1.0
